requirements.txt
torch
torchvision
torchaudio
numpy
pillow